Improve tracking error logging message
The error message logged when tracking posting fails now makes it more
obvious that it is related to DBT's internal anonymous usage
statistics, and explains how to disable it.

As tracking failure doesn't real concern most users, we also reduce the
logging level to "DEBUG" so it doesn't appear during normal use.

Issue #5197

Signed-off-by: NicolasPA <22459427+NicolasPA@users.noreply.github.com>