refactor: Simplify configuration and set_connection_handle signature
With the objective of simplifying the configuration exposed to a user,
and empowering dbt adapter manitainers to define their own custom
behavior, several retry configuration options have been removed,
leaving only retries.

Moreover, set_connection_handle now takes an inclusive list of
exceptions to retry on, and an exclusive list of exceptions to not
retry on. Adapter maintainers should customize these to a default that
works for their specific database.

Finally, the timeout argument of set_connection_handle can now take a
callable that returns an int or float to pass to time.sleep. This
allows adapter maintainers to implement more complicated backoff algorithms.