refactor: Simplify retry connection with recursive method
This refactores set_connection_handle to be now retry_connection and
be implemented with a recursive approach instead of a while loop.

acquire_handle is now connect and is passed as an argument to
retry_connection, instead of being a method in Credentials.

Moreover, several optional arguments were removed: we now only take an
iterable of exceptions to catch and retry on. This significantly
simplifies things.

Tests have been updated accordingly.