add utility functions to EventManager for explicitly adding callbacks
Technically these aren't necessary in their current state. We could instead
have people do `<InstantiatedEventManager>.callbacks.extend(...)` directly.
However, it's not hard to imagine a world wherein extra things need to take
place when a callback is added. Thus abstracting to these utility methods
now means that as the implementation of how callbacks are actually added
changes, the invocation to do so can stay the same.