First pass at refactoring MetricNode definition to satisfy DSI Metric protocol
This isn't the "final" implementation/refactor of the Metric node for this branch,
the first of a few. This just gets us to a place where the MetricNode satisfies
the DSI Metric protocol even if some things are left unimplemented. Additioanlly,
this does not include the changes necessary for parsing