Update the `SemanticModel` node to match DSI protocol
A lot of the additions are helper functions which we don't actually
use on core. This is a known issue. We're in the process of removing
a fair number of them from the DSI protocol spec. However, in the mean
time we need to implement them to satisfy the protocol unfortunately.