Update typing-extensions version to >= 4.4
With the addition of dbt-semantic-interfaces as a package dependency
for dbt-core we added an implicit requirement that the local install
of typing-extensions be for a version >= 4.4. This means, effective
with dbt-core 1.6, our version minimum for typing-extensions has been
updated to 4.4.

This change reflects the reality imposed on us by our dependency hierarchy.
Happily, mashumaro previously boosted its base version dependency to 4.1
some time ago, and effective with major version 4 the typing-extensions
package maintainers have committed to strictly following SemVer, so this
should be a low risk minimum version change.