Fix `test_graph.py` config creation process to ensure flags exist on config
Previously in our `test_graph.py` unit tests we were setting the flags global,
but not actually adding them to the config. The config, without the flags,
would then get passed to the manifest loader and other things. Those down stream
classes/functions would then look to the config for the flags and not find them.
This change ensures that the flags get applied to the config that is being used
in down stream operations during our unit tests.